Install Flask-Caching.py
To implement Redis caching for the information displayed in the table on page 2, we'll use the Flask-Caching extension along with Redis as the caching backend. This will help reduce the number of database read requests by caching the data retrieved from the database.

First, ensure you have Redis installed and running on your system.

pip install Flask-Caching


-> We import the Cache extension from Flask-Caching.

-> We configure the caching backend to use Redis (CACHE_TYPE = 'redis') and specify the Redis URL (CACHE_REDIS_URL = 'redis://localhost:6379/0').

->We create a cache instance with Cache(app).
In the submissions route, we first check if the submissions data is available in the cache. If it's not, we fetch the data from the database, cache it, and set a timeout for the cache (in this case, 60 seconds).

->We delete the cache key 'submissions' after a new submission is made to ensure that the cache is refreshed.